+++ app/build.gradle
@@ -8,8 +8,8 @@ android {
         minSdkVersion 15
         targetSdkVersion 29
 
-        versionCode 32
-        versionName "2.0.14"
+        versionCode 34
+        versionName "2.0.16"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         setProperty("archivesBaseName", project.getParent().name+"-"+versionName)
     }
