--- app/src/main/java/io/github/chronosx88/yggdrasil/models/NodeInfo.kt
@@ -1,13 +1,0 @@
-package io.github.chronosx88.yggdrasil.models
-
-class NodeInfo {
-
-    constructor(key: String, value: String){
-        this.key = key
-        this.value = value
-    }
-
-    var key: String
-    var value: String
-
-}
