--- app/src/main/java/io/github/chronosx88/yggdrasil/models/Status.kt
@@ -1,5 +1,0 @@
-package io.github.chronosx88.yggdrasil.models
-
-class Status {
-    var up: Boolean = false
-}
