+++ app/src/main/java/org/yggdrasil/app/crispa/models/NodeInfo.kt
@@ -1,0 +1,13 @@
+package org.yggdrasil.app.crispa.models
+
+class NodeInfo {
+
+    constructor(key: String, value: String){
+        this.key = key
+        this.value = value
+    }
+
+    var key: String
+    var value: String
+
+}
