+++ app/src/main/java/org/yggdrasil/app/crispa/models/Status.kt
@@ -1,0 +1,5 @@
+package org.yggdrasil.app.crispa.models
+
+class Status {
+    var up: Boolean = false
+}
