+++ app/src/main/java/org/yggdrasil/app/crispa/models/config/SwitchOptions.kt
@@ -1,0 +1,7 @@
+package org.yggdrasil.app.crispa.models.config
+
+import com.google.gson.annotations.SerializedName
+
+data class SwitchOptions (
+	@SerializedName("maxTotalQueueSize") var maxTotalQueueSize : Int // Maximum size of all switch queues combined (in bytes).
+)
